redistribution of the abc factory method
